+++ ChangeLog.md
@@ -3,6 +3,11 @@ XP Framework Core ChangeLog
 
 ## ?.?.? / ????-??-??
 
+### Bugfixes
+
+* Fixed `math.BigNum` division on HHVM cutting off below default precision
+  (@thekid)
+
 ## 6.3.1 / 2015-06-13
 
 ### Heads up!